refactor: moved limit and receipts too
