+++ op-service/client/eth_client_test.go
@@ -1,4 +1,4 @@
-package sources
+package client
 
 import (
 	"context"
